feat: create DetailServiceList (hardcode)
